System state in System info
You weren't able to see the current status of the NAS. For example if the disks are synchronizing. Now you can see if the NAS is Ready or is in a warning state. 
Also the command used to get the SNMP value can be made cleaner. Less regex and less output from snmpget.